keep garbage collection logic separate from index
