+++ quickwit/quickwit-index-management/Cargo.toml
@@ -21,6 +21,7 @@ serde = { workspace = true }
 serde_json = "1.0"
 tantivy = { workspace = true }
 tempfile = { workspace = true }
+time = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
 tokio-stream = { workspace = true }
